django/requirements.txt
Django==2.0.1
djangorestframework==3.7.7
psycopg2>=2.7.5,<2.8.0
Pillow>=5.3.0,<5.4.0
flake8>=3.6.0,<3.7.0